Implement and test the Create() method
I am still figuring out how to test.
An integration test with a real DB would be best...

It does not return duplicate ID error yet, because I have no idea of the
real code returned by SQLite in such a case.
This drops the dependency on go-sqlite3 (which should not be needed).