Add an integration test
It adds back the dependency to go-sqlite3. Maybe we should not depend on
it but for now, it's better to move on and advance the project.

The integration test just creates an in-memory SQLite database, creates
the sessions table and adds a session. Later it will do the other
operations in sessionup interface.